refactor(dat)!: change the loading mechanism
Previously, any Dat could be created dynamically by using Dat.load. The
spec would be inspected for a dat.class, and any subclass of Dat that
matched dat.class would be used to instantiate the loaded object. This
has the drawback of dynamically creating a class of unknown type until
runtime, and that the class must be defined ahead of time for the logic
to find it as a subclass of Dat automatically.

Now you can still load any Dat with Dat.load, and it'll be guaranteed
to be a dat (as long as it's valid). But you can also use
DatSubclass.load, and the returned instance will be of type DatSubclass,
which addresses the second problem described above.